separate move group runner to scheduler & dispatcher
